Splitting the logic and runnable
