+++ TheMindGameJava/src/TheMindGameLogicResult.java
@@ -1,13 +1,18 @@
 import ReadModel.Card;
 
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-public class TheMindGameResult {
+public class TheMindGameLogicResult {
     private final int failedGames;
     private final Map<Integer, List<Card>> result;
 
-    public TheMindGameResult(int failedGames, Map<Integer, List<Card>> result) {
+    public static TheMindGameLogicResult createEmpty() {
+        return new TheMindGameLogicResult(0, new HashMap<>());
+    }
+
+    public TheMindGameLogicResult(int failedGames, Map<Integer, List<Card>> result) {
         this.failedGames = failedGames;
         this.result = result;
     }
@@ -15,5 +20,9 @@ public class TheMindGameResult {
     @Override
     public String toString() {
         return "failedGames=" + failedGames + ", " + result;
+    }
+
+    public boolean isEmpty() {
+        return result.isEmpty();
     }
 }
